+++ src/main/java/Main.java
@@ -3,7 +3,6 @@ import client.NanoClient;
 import com.jagrosh.jdautilities.command.CommandClient;
 import com.jagrosh.jdautilities.command.CommandClientBuilder;
 import com.jagrosh.jdautilities.commons.waiter.EventWaiter;
-import command.*;
 import command.general.CustomPrefixCommand;
 import command.general.HelpCommand;
 import command.general.InviteCommand;
@@ -11,6 +10,7 @@ import command.general.VoteCommand;
 import command.history.UserHistoryCommand;
 import command.music.*;
 import command.owner.ChangePresenceCommand;
+import command.owner.PremiumUserCommand;
 import listener.MemberVoiceListener;
 import net.dv8tion.jda.api.JDA;
 import net.dv8tion.jda.api.JDABuilder;
